% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countStart.R
\name{countStart}
\alias{countStart}
\title{Count the number of categorical observations starting with a specific letter.}
\usage{
countStart(df, categoricalCol, letter)
}
\arguments{
\item{df}{Data or data frame.}

\item{categoricalCol}{A column containing categorical variables.}

\item{letter}{A letter to filter through the starting letter of categorical variables.}
}
\value{
The count of categorical observations starting with the specified letter.
}
\description{
This function looks down a categorical column of a data set and returns the count of categorical observations
that start with a given letter.
}
\examples{
countStart(palmerpenguins::penguins, species, "G")

}
