% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamesCount.R
\name{getNamesCount}
\alias{getNamesCount}
\title{Count Names That Fall Within a Numerical Value Range}
\usage{
getNamesCount(df, categoricalCol, numericalCol, numericalValue)
}
\arguments{
\item{df}{Data or data frame}

\item{categoricalCol}{A column containing categorical variables.}

\item{numericalCol}{A column contain numerical values.}

\item{numericalValue}{A number from the numerical values column.}
}
\value{
This function returns a numerical output, from the row counts after filtering.
}
\description{
The getNamesCount function can be used to count the number of observations in a categorical column that meet the value criteria specified from a numerical column.
}
\examples{
 dfSample <- data.frame(
fruit = c("apple", "banana", "cherry", "coconut", "dragonfruit"),
numFruit = c(17, 90, 35, 49, 102))
getNamesCount(dfSample, fruit, numFruit, 50)

getNamesCount(palmerpenguins::penguins, species, bill_length_mm, 43.92)

}
