% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNamesUnique.R
\name{getNamesUnique}
\alias{getNamesUnique}
\title{Display all Unique Names within a Categorical Column}
\usage{
getNamesUnique(df, categoricalCol)
}
\arguments{
\item{df}{Data or data frame}

\item{categoricalCol}{A column containing categorical variables.}
}
\value{
This function returns a list of unique categorical variables.
}
\description{
The getNamesUnique function can be used used to display the unique categorical variables present in a specified column of a data frame.
}
\examples{
 dfSample <- data.frame(
fruit = c("apple", "banana", "cherry", "coconut", "dragonfruit"),
numFruit = c(17, 90, 35, 49, 102))
getNamesUnique(dfSample, fruit)

getNamesUnique(palmerpenguins::penguins, species)

}
